dashboard/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.2.0
pydantic>=2.10.0
openai>=1.60.0
anthropic>=0.42.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.28.0
